Add support to Mooncake
I add Radio buttons to let users to choose if they are working with
Global Azure or Mooncake. By default the Global Azure option is checked.

And the Add AzureEnvironment varible to differentiate it. This variable
will be used when creating service client since the service endpoints
are different from Global Azure and Mooncake.